Remove DataTable objects and merge with DataPresenter
