+++ qt/scientific_interfaces/Indirect/ConvFitDataView.h
@@ -6,34 +6,30 @@
 // SPDX - License - Identifier: GPL - 3.0 +
 #pragma once
 
-#include "ConvFitModel.h"
-#include "IndirectFitDataTablePresenter.h"
+#include "IndirectFitDataView.h"
 
-#include <QTableWidget>
+#include "DllConfig.h"
 
-#include <cstddef>
-#include <unordered_map>
+#include <QTabWidget>
 
 namespace MantidQt {
 namespace CustomInterfaces {
 namespace IDA {
-
 /**
-  Presenter for a table of convolution fitting data.
+Presenter for a table of convolution fitting data.
 */
-class DLLExport ConvFitDataTablePresenter : public IndirectFitDataTablePresenter {
+class MANTIDQT_INDIRECT_DLL ConvFitDataView : public IndirectFitDataView {
   Q_OBJECT
 public:
-  ConvFitDataTablePresenter(ConvFitModel *model, QTableWidget *dataTable);
-
-protected:
-  void addTableEntry(FitDomainIndex row) override;
-
-private:
+  ConvFitDataView(QWidget *parent = nullptr);
+  void addTableEntry(size_t row, FitDataRow newRow) override;
   int workspaceIndexColumn() const override;
   int startXColumn() const override;
   int endXColumn() const override;
   int excludeColumn() const override;
+
+protected:
+  ConvFitDataView(const QStringList &headers, QWidget *parent = nullptr);
 };
 
 } // namespace IDA
