+++ qt/scientific_interfaces/Indirect/IIndirectFitDataView.h
@@ -7,6 +7,7 @@
 #pragma once
 
 #include "DllConfig.h"
+#include "MantidQtWidgets/Common/IndexTypes.h"
 #include "MantidQtWidgets/Common/UserInputValidator.h"
 
 #include <QObject>
@@ -16,6 +17,16 @@
 namespace MantidQt {
 namespace CustomInterfaces {
 namespace IDA {
+
+struct FitDataRow {
+  std::string name;
+  std::string exclude;
+  size_t workspaceIndex;
+  int startX;
+  int endX;
+  std::string resolution;
+  std::string parameter;
+};
 
 class MANTIDQT_INDIRECT_DLL IIndirectFitDataView : public QTabWidget {
   Q_OBJECT
@@ -27,6 +38,14 @@ public:
   virtual QTableWidget *getDataTable() const = 0;
 
   virtual UserInputValidator &validate(UserInputValidator &validator) = 0;
+  virtual void addTableEntry(size_t row, FitDataRow newRow) = 0;
+  virtual int workspaceIndexColumn() const = 0;
+  virtual int startXColumn() const = 0;
+  virtual int endXColumn() const = 0;
+  virtual int excludeColumn() const = 0;
+  virtual void clearTable() = 0;
+  virtual QString getText(int row, int column) const = 0;
+  virtual QModelIndexList getSelectedIndexes() const = 0;
 
 public slots:
   virtual void displayWarning(std::string const &warning) = 0;
