+++ qt/scientific_interfaces/Indirect/IndirectDataAnalysisIqtFitTab.cpp
@@ -41,7 +41,8 @@ IndirectDataAnalysisIqtFitTab::IndirectDataAnalysisIqtFitTab(QWidget *parent)
     : IndirectFitAnalysisTab(new IqtFitModel, parent), m_uiForm(new Ui::IndirectFitTab) {
   m_uiForm->setupUi(parent);
   m_iqtFittingModel = dynamic_cast<IqtFitModel *>(getFittingModel());
-  setFitDataPresenter(std::make_unique<IndirectFitDataPresenter>(m_iqtFittingModel, m_uiForm->dockArea->m_fitDataView));
+  setFitDataPresenter(std::make_unique<IndirectFitDataPresenter>(m_iqtFittingModel->getFitDataModel(),
+                                                                 m_uiForm->dockArea->m_fitDataView));
   setPlotView(m_uiForm->dockArea->m_fitPlotView);
   setOutputOptionsView(m_uiForm->ovOutputOptionsView);
   auto templateBrowser = new IqtTemplateBrowser;
