+++ qt/scientific_interfaces/Indirect/IndirectFitDataModel.cpp
@@ -25,9 +25,6 @@ bool equivalentWorkspaces(const Mantid::API::MatrixWorkspace_const_sptr &lhs,
   return lhs->getName() == rhs->getName();
 }
 
-bool doesExistInADS(std::string const &workspaceName) {
-  return Mantid::API::AnalysisDataService::Instance().doesExist(workspaceName);
-}
 } // namespace
 
 namespace MantidQt {
@@ -35,52 +32,46 @@ namespace CustomInterfaces {
 namespace IDA {
 
 IndirectFitDataModel::IndirectFitDataModel()
-    : m_fittingDataSingle(std::make_unique<std::vector<IndirectFitData>>()),
-      m_resolutionsSingle(std::make_unique<std::vector<std::weak_ptr<Mantid::API::MatrixWorkspace>>>()),
-      m_fittingDataMultiple(std::make_unique<std::vector<IndirectFitData>>()),
-      m_resolutionsMultiple(std::make_unique<std::vector<std::weak_ptr<Mantid::API::MatrixWorkspace>>>()) {
-
-  m_fittingData = m_fittingDataSingle.get();
-  m_resolutions = m_resolutionsSingle.get();
-}
+    : m_fittingData(std::make_unique<std::vector<IndirectFitData>>()),
+      m_resolutions(std::make_unique<std::vector<std::weak_ptr<Mantid::API::MatrixWorkspace>>>()),
+      m_adsInstance(Mantid::API::AnalysisDataService::Instance()) {}
 
 bool IndirectFitDataModel::hasWorkspace(std::string const &workspaceName) const {
   auto const names = getWorkspaceNames();
-  auto const iter = std::find(names.begin(), names.end(), workspaceName);
-  return iter != names.end();
+  auto const iter = std::find(names.cbegin(), names.cend(), workspaceName);
+  return iter != names.cend();
 }
 
-Mantid::API::MatrixWorkspace_sptr IndirectFitDataModel::getWorkspace(TableDatasetIndex index) const {
-  if (index < m_fittingData->size())
-    return m_fittingData->at(index.value).workspace();
+Mantid::API::MatrixWorkspace_sptr IndirectFitDataModel::getWorkspace(WorkspaceID workspaceID) const {
+  if (workspaceID < m_fittingData->size())
+    return m_fittingData->at(workspaceID.value).workspace();
   return nullptr;
 }
 
-FunctionModelSpectra IndirectFitDataModel::getSpectra(TableDatasetIndex index) const {
-  if (index < m_fittingData->size())
-    return m_fittingData->at(index.value).spectra();
+FunctionModelSpectra IndirectFitDataModel::getSpectra(WorkspaceID workspaceID) const {
+  if (workspaceID < m_fittingData->size())
+    return m_fittingData->at(workspaceID.value).spectra();
   return FunctionModelSpectra("");
 }
 
-TableDatasetIndex IndirectFitDataModel::numberOfWorkspaces() const { return TableDatasetIndex{m_fittingData->size()}; }
+WorkspaceID IndirectFitDataModel::getNumberOfWorkspaces() const { return WorkspaceID{m_fittingData->size()}; }
 
-bool IndirectFitDataModel::isMultiFit() const { return numberOfWorkspaces().value > 1; }
+bool IndirectFitDataModel::isMultiFit() const { return getNumberOfWorkspaces().value > 1; }
 
-size_t IndirectFitDataModel::getNumberOfSpectra(TableDatasetIndex index) const {
-  if (index < m_fittingData->size())
-    return m_fittingData->at(index.value).numberOfSpectra().value;
+size_t IndirectFitDataModel::getNumberOfSpectra(WorkspaceID workspaceID) const {
+  if (workspaceID < m_fittingData->size())
+    return m_fittingData->at(workspaceID.value).numberOfSpectra().value;
   else
     throw std::runtime_error("Cannot find the number of spectra for a workspace: the workspace "
                              "index provided is too large.");
-  return 1;
 }
 
 size_t IndirectFitDataModel::getNumberOfDomains() const {
-  size_t sum{0};
-  for (auto fittingData : *m_fittingData) {
-    sum += fittingData.numberOfSpectra().value;
-  }
-  return sum;
+  size_t init(0);
+  auto const getNumberOfSpectra = [](size_t sum, auto const &fittingData) {
+    return sum + fittingData.numberOfSpectra().value;
+  };
+  return std::accumulate(m_fittingData->cbegin(), m_fittingData->cend(), init, getNumberOfSpectra);
 }
 
 std::vector<double> IndirectFitDataModel::getQValuesForData() const {
@@ -95,8 +86,8 @@ std::vector<double> IndirectFitDataModel::getQValuesForData() const {
 std::vector<std::pair<std::string, size_t>> IndirectFitDataModel::getResolutionsForFit() const {
   std::vector<std::pair<std::string, size_t>> resolutionVector;
   for (size_t index = 0; index < m_resolutions->size(); ++index) {
-    auto resolutionWorkspace = m_resolutions->at(index).lock();
-    auto spectra = getSpectra(TableDatasetIndex{index});
+    const auto resolutionWorkspace = m_resolutions->at(index).lock();
+    const auto spectra = getSpectra(WorkspaceID{index});
     if (!resolutionWorkspace) {
       for (auto &spectraIndex : spectra) {
         resolutionVector.emplace_back("", spectraIndex.value);
@@ -104,44 +95,45 @@ std::vector<std::pair<std::string, size_t>> IndirectFitDataModel::getResolutions
       continue;
     }
 
-    auto singleSpectraResolution = m_resolutions->at(index).lock()->getNumberHistograms() == 1;
+    const auto singleSpectraResolution = m_resolutions->at(index).lock()->getNumberHistograms() == 1;
     for (auto &spectraIndex : spectra) {
-      auto resolutionIndex = singleSpectraResolution ? 0 : spectraIndex.value;
+      const auto resolutionIndex = singleSpectraResolution ? 0 : spectraIndex.value;
       resolutionVector.emplace_back(m_resolutions->at(index).lock()->getName(), resolutionIndex);
     }
   }
   return resolutionVector;
 }
 
-void IndirectFitDataModel::setResolution(const std::string &name, TableDatasetIndex index) {
-  if (!name.empty() && doesExistInADS(name)) {
-    auto resolution = Mantid::API::AnalysisDataService::Instance().retrieveWS<Mantid::API::MatrixWorkspace>(name);
-    if (m_resolutions->size() > index.value) {
-      m_resolutions->at(index.value) = resolution;
-    } else if (m_resolutions->size() == index.value) {
+void IndirectFitDataModel::setResolution(const std::string &name, WorkspaceID workspaceID) {
+  if (!name.empty() && m_adsInstance.doesExist(name)) {
+    const auto resolution = m_adsInstance.retrieveWS<Mantid::API::MatrixWorkspace>(name);
+    if (m_resolutions->size() > workspaceID.value) {
+      m_resolutions->at(workspaceID.value) = resolution;
+    } else if (m_resolutions->size() == workspaceID.value) {
       m_resolutions->emplace_back(resolution);
     } else {
-      throw std::out_of_range("Provided resolution index '" + std::to_string(index.value) + "' was out of range.");
+      throw std::out_of_range("Provided resolution index '" + std::to_string(workspaceID.value) +
+                              "' was out of range.");
     }
   } else {
     throw std::runtime_error("A valid resolution file needs to be selected.");
   }
 }
 
-void IndirectFitDataModel::setSpectra(const std::string &spectra, TableDatasetIndex dataIndex) {
-  setSpectra(FunctionModelSpectra(spectra), dataIndex);
+void IndirectFitDataModel::setSpectra(const std::string &spectra, WorkspaceID workspaceID) {
+  setSpectra(FunctionModelSpectra(spectra), workspaceID);
 }
 
-void IndirectFitDataModel::setSpectra(FunctionModelSpectra &&spectra, TableDatasetIndex dataIndex) {
+void IndirectFitDataModel::setSpectra(FunctionModelSpectra &&spectra, WorkspaceID workspaceID) {
   if (m_fittingData->empty())
     return;
-  m_fittingData->at(dataIndex.value).setSpectra(std::forward<FunctionModelSpectra>(spectra));
+  m_fittingData->at(workspaceID.value).setSpectra(std::forward<FunctionModelSpectra>(spectra));
 }
 
-void IndirectFitDataModel::setSpectra(const FunctionModelSpectra &spectra, TableDatasetIndex dataIndex) {
+void IndirectFitDataModel::setSpectra(const FunctionModelSpectra &spectra, WorkspaceID workspaceID) {
   if (m_fittingData->empty())
     return;
-  m_fittingData->at(dataIndex.value).setSpectra(spectra);
+  m_fittingData->at(workspaceID.value).setSpectra(spectra);
 }
 
 std::vector<std::string> IndirectFitDataModel::getWorkspaceNames() const {
@@ -153,14 +145,14 @@ std::vector<std::string> IndirectFitDataModel::getWorkspaceNames() const {
 }
 
 void IndirectFitDataModel::addWorkspace(const std::string &workspaceName) {
-  auto ws = Mantid::API::AnalysisDataService::Instance().retrieveWS<Mantid::API::MatrixWorkspace>(workspaceName);
+  auto ws = m_adsInstance.retrieveWS<Mantid::API::MatrixWorkspace>(workspaceName);
   addWorkspace(ws, FunctionModelSpectra(WorkspaceIndex{0}, WorkspaceIndex{ws->getNumberHistograms() - 1}));
 }
 
 void IndirectFitDataModel::addWorkspace(const std::string &workspaceName, const std::string &spectra) {
   if (spectra.empty())
     throw std::runtime_error("Fitting Data must consist of one or more spectra.");
-  if (workspaceName.empty() || !Mantid::API::AnalysisDataService::Instance().doesExist(workspaceName))
+  if (workspaceName.empty() || !m_adsInstance.doesExist(workspaceName))
     throw std::runtime_error("A valid sample file needs to be selected.");
 
   try {
@@ -171,16 +163,16 @@ void IndirectFitDataModel::addWorkspace(const std::string &workspaceName, const 
 }
 
 void IndirectFitDataModel::addWorkspace(const std::string &workspaceName, const FunctionModelSpectra &spectra) {
-  auto ws = Mantid::API::AnalysisDataService::Instance().retrieveWS<Mantid::API::MatrixWorkspace>(workspaceName);
+  auto ws = m_adsInstance.retrieveWS<Mantid::API::MatrixWorkspace>(workspaceName);
   addWorkspace(ws, spectra);
 }
 
 void IndirectFitDataModel::addWorkspace(Mantid::API::MatrixWorkspace_sptr workspace,
                                         const FunctionModelSpectra &spectra) {
   if (!m_fittingData->empty()) {
-    for (auto i : *m_fittingData) {
-      if (equivalentWorkspaces(workspace, i.workspace())) {
-        i.combine(IndirectFitData(workspace, spectra));
+    for (auto &fitData : *m_fittingData) {
+      if (equivalentWorkspaces(workspace, fitData.workspace())) {
+        fitData.combine(IndirectFitData(workspace, spectra));
         return;
       }
     }
@@ -188,10 +180,10 @@ void IndirectFitDataModel::addWorkspace(Mantid::API::MatrixWorkspace_sptr worksp
   addNewWorkspace(workspace, spectra);
 }
 
-FitDomainIndex IndirectFitDataModel::getDomainIndex(TableDatasetIndex dataIndex, WorkspaceIndex spectrum) const {
+FitDomainIndex IndirectFitDataModel::getDomainIndex(WorkspaceID workspaceID, WorkspaceIndex spectrum) const {
   FitDomainIndex index{0};
   for (size_t iws = 0; iws < m_fittingData->size(); ++iws) {
-    if (iws < dataIndex.value) {
+    if (iws < workspaceID.value) {
       index += getNumberOfSpectra(iws);
     } else {
       auto const spectra = getSpectra(iws);
@@ -209,50 +201,50 @@ FitDomainIndex IndirectFitDataModel::getDomainIndex(TableDatasetIndex dataIndex,
 
 void IndirectFitDataModel::clear() { m_fittingData->clear(); }
 
-std::pair<double, double> IndirectFitDataModel::getFittingRange(TableDatasetIndex dataIndex,
+std::pair<double, double> IndirectFitDataModel::getFittingRange(WorkspaceID workspaceID,
                                                                 WorkspaceIndex spectrum) const {
-  if (dataIndex.value < m_fittingData->size() && !m_fittingData->at(dataIndex.value).zeroSpectra()) {
-    return m_fittingData->at(dataIndex.value).getRange(spectrum);
+  if (workspaceID.value < m_fittingData->size() && !m_fittingData->at(workspaceID.value).zeroSpectra()) {
+    return m_fittingData->at(workspaceID.value).getRange(spectrum);
   }
   return std::make_pair(0., 0.);
 }
 
-std::string IndirectFitDataModel::getExcludeRegion(TableDatasetIndex dataIndex, WorkspaceIndex spectrum) const {
-  if (dataIndex.value < m_fittingData->size() && !m_fittingData->at(dataIndex.value).zeroSpectra()) {
-    return m_fittingData->at(dataIndex.value).getExcludeRegion(spectrum);
+std::string IndirectFitDataModel::getExcludeRegion(WorkspaceID workspaceID, WorkspaceIndex spectrum) const {
+  if (workspaceID.value < m_fittingData->size() && !m_fittingData->at(workspaceID.value).zeroSpectra()) {
+    return m_fittingData->at(workspaceID.value).getExcludeRegion(spectrum);
   }
   return "";
 }
 
-void IndirectFitDataModel::setStartX(double startX, TableDatasetIndex dataIndex, WorkspaceIndex spectrum) {
+void IndirectFitDataModel::setStartX(double startX, WorkspaceID workspaceID, WorkspaceIndex spectrum) {
   if (m_fittingData->empty())
     return;
-  m_fittingData->at(dataIndex.value).setStartX(startX, spectrum);
+  m_fittingData->at(workspaceID.value).setStartX(startX, spectrum);
 }
 
-void IndirectFitDataModel::setStartX(double startX, TableDatasetIndex dataIndex) {
+void IndirectFitDataModel::setStartX(double startX, WorkspaceID workspaceID) {
   if (m_fittingData->empty())
     return;
-  m_fittingData->at(dataIndex.value).setStartX(startX);
+  m_fittingData->at(workspaceID.value).setStartX(startX);
 }
 
-void IndirectFitDataModel::setEndX(double endX, TableDatasetIndex dataIndex, WorkspaceIndex spectrum) {
+void IndirectFitDataModel::setEndX(double endX, WorkspaceID workspaceID, WorkspaceIndex spectrum) {
   if (m_fittingData->empty())
     return;
-  m_fittingData->at(dataIndex.value).setEndX(endX, spectrum);
+  m_fittingData->at(workspaceID.value).setEndX(endX, spectrum);
 }
 
-void IndirectFitDataModel::setEndX(double endX, TableDatasetIndex dataIndex) {
+void IndirectFitDataModel::setEndX(double endX, WorkspaceID workspaceID) {
   if (m_fittingData->empty())
     return;
-  m_fittingData->at(dataIndex.value).setEndX(endX);
+  m_fittingData->at(workspaceID.value).setEndX(endX);
 }
 
-void IndirectFitDataModel::setExcludeRegion(const std::string &exclude, TableDatasetIndex dataIndex,
+void IndirectFitDataModel::setExcludeRegion(const std::string &exclude, WorkspaceID workspaceID,
                                             WorkspaceIndex spectrum) {
   if (m_fittingData->empty())
     return;
-  m_fittingData->at(dataIndex.value).setExcludeRegionString(exclude, spectrum);
+  m_fittingData->at(workspaceID.value).setExcludeRegionString(exclude, spectrum);
 }
 
 void IndirectFitDataModel::addNewWorkspace(const Mantid::API::MatrixWorkspace_sptr &workspace,
@@ -260,9 +252,9 @@ void IndirectFitDataModel::addNewWorkspace(const Mantid::API::MatrixWorkspace_sp
   m_fittingData->emplace_back(workspace, spectra);
 }
 
-void IndirectFitDataModel::removeWorkspace(TableDatasetIndex index) {
-  if (index < m_fittingData->size()) {
-    m_fittingData->erase(m_fittingData->begin() + index.value);
+void IndirectFitDataModel::removeWorkspace(WorkspaceID workspaceID) {
+  if (workspaceID < m_fittingData->size()) {
+    m_fittingData->erase(m_fittingData->begin() + workspaceID.value);
   } else {
     throw std::runtime_error("Attempting to remove non-existent workspace.");
   }
@@ -279,20 +271,10 @@ void IndirectFitDataModel::removeDataByIndex(FitDomainIndex fitDomainIndex) {
   }
 }
 
-void IndirectFitDataModel::switchToSingleInputMode() {
-  m_fittingData = m_fittingDataSingle.get();
-  m_resolutions = m_resolutionsSingle.get();
-}
-
-void IndirectFitDataModel::switchToMultipleInputMode() {
-  m_fittingData = m_fittingDataMultiple.get();
-  m_resolutions = m_resolutionsMultiple.get();
-}
-
-std::vector<double> IndirectFitDataModel::getExcludeRegionVector(TableDatasetIndex dataIndex,
-                                                                 WorkspaceIndex index) const {
-  auto fitData = m_fittingData->at(dataIndex.value);
-  return fitData.excludeRegionsVector(index);
+std::vector<double> IndirectFitDataModel::getExcludeRegionVector(WorkspaceID workspaceID,
+                                                                 WorkspaceIndex spectrum) const {
+  auto fitData = m_fittingData->at(workspaceID.value);
+  return fitData.excludeRegionsVector(spectrum);
 }
 
 Mantid::API::MatrixWorkspace_sptr IndirectFitDataModel::getWorkspace(FitDomainIndex index) const {
@@ -323,18 +305,18 @@ std::string IndirectFitDataModel::getExcludeRegion(FitDomainIndex index) const {
 void IndirectFitDataModel::setExcludeRegion(const std::string &exclude, FitDomainIndex index) {
   if (m_fittingData->empty())
     return;
-  auto subIndices = getSubIndices(index);
+  const auto subIndices = getSubIndices(index);
   m_fittingData->at(subIndices.first.value).setExcludeRegionString(exclude, subIndices.second);
 }
 
-std::pair<TableDatasetIndex, WorkspaceIndex> IndirectFitDataModel::getSubIndices(FitDomainIndex index) const {
+std::pair<WorkspaceID, WorkspaceIndex> IndirectFitDataModel::getSubIndices(FitDomainIndex index) const {
   size_t sum{0};
-  for (size_t datasetIndex = 0; datasetIndex < m_fittingData->size(); datasetIndex++) {
-    for (size_t workspaceIndex = 0; workspaceIndex < m_fittingData->at(datasetIndex).spectra().size().value;
+  for (size_t workspaceID = 0; workspaceID < m_fittingData->size(); workspaceID++) {
+    for (size_t workspaceIndex = 0; workspaceIndex < m_fittingData->at(workspaceID).spectra().size().value;
          workspaceIndex++) {
       if (sum == index.value) {
-        WorkspaceIndex spectraIndex = m_fittingData->at(datasetIndex).spectra()[FitDomainIndex{workspaceIndex}];
-        return std::make_pair(TableDatasetIndex{datasetIndex}, spectraIndex);
+        WorkspaceIndex spectrum = m_fittingData->at(workspaceID).spectra()[FitDomainIndex{workspaceIndex}];
+        return std::make_pair(WorkspaceID{workspaceID}, spectrum);
       }
       sum++;
     }
