+++ qt/scientific_interfaces/Indirect/IndirectFitDataPresenter.cpp
@@ -10,33 +10,38 @@
 
 #include "IndirectAddWorkspaceDialog.h"
 
+namespace {
+
+QString makeNumber(double d) { return QString::number(d, 'g', 16); }
+
+class ScopedFalse {
+  bool &m_ref;
+  bool m_oldValue;
+
+public:
+  // this sets the input bool to false whilst this object is in scope and then
+  // resets it to its old value when this object drops out of scope.
+  explicit ScopedFalse(bool &variable) : m_ref(variable), m_oldValue(variable) { m_ref = false; }
+  ~ScopedFalse() { m_ref = m_oldValue; }
+};
+} // namespace
+
 namespace MantidQt {
 namespace CustomInterfaces {
 namespace IDA {
 
-IndirectFitDataPresenter::IndirectFitDataPresenter(IIndirectFittingModel *model, IIndirectFitDataView *view)
-    : IndirectFitDataPresenter(
-          model, view,
-          std::make_unique<IndirectFitDataTablePresenter>(model->getFitDataModel(), view->getDataTable())) {}
-
-IndirectFitDataPresenter::IndirectFitDataPresenter(IIndirectFittingModel *model, IIndirectFitDataView *view,
-                                                   std::unique_ptr<IndirectFitDataTablePresenter> tablePresenter)
-    : m_model(model), m_view(view), m_tablePresenter(std::move(tablePresenter)) {
+IndirectFitDataPresenter::IndirectFitDataPresenter(IIndirectFitDataModel *model, IIndirectFitDataView *view)
+    : m_model(model), m_view(view) {
   observeReplace(true);
 
   connect(m_view, SIGNAL(addClicked()), this, SIGNAL(requestedAddWorkspaceDialog()));
   connect(m_view, SIGNAL(addClicked()), this, SLOT(showAddWorkspaceDialog()));
 
-  connect(m_view, SIGNAL(removeClicked()), m_tablePresenter.get(), SLOT(removeSelectedData()));
+  connect(m_view, SIGNAL(removeClicked()), this, SLOT(removeSelectedData()));
   connect(m_view, SIGNAL(removeClicked()), this, SIGNAL(dataRemoved()));
   connect(m_view, SIGNAL(removeClicked()), this, SIGNAL(dataChanged()));
   connect(m_view, SIGNAL(startXChanged(double)), this, SIGNAL(startXChanged(double)));
   connect(m_view, SIGNAL(endXChanged(double)), this, SIGNAL(endXChanged(double)));
-
-  connect(m_tablePresenter.get(), SIGNAL(startXChanged(double, WorkspaceID, WorkspaceIndex)), this,
-          SIGNAL(startXChanged(double, WorkspaceID, WorkspaceIndex)));
-  connect(m_tablePresenter.get(), SIGNAL(endXChanged(double, WorkspaceID, WorkspaceIndex)), this,
-          SIGNAL(endXChanged(double, WorkspaceID, WorkspaceIndex)));
 }
 
 IndirectFitDataPresenter::~IndirectFitDataPresenter() { observeReplace(false); }
@@ -62,13 +67,6 @@ QStringList IndirectFitDataPresenter::getSampleFBSuffices() const { return m_fbS
 QStringList IndirectFitDataPresenter::getResolutionWSSuffices() const { return m_wsResolutionSuffixes; }
 
 QStringList IndirectFitDataPresenter::getResolutionFBSuffices() const { return m_fbResolutionSuffixes; }
-
-void IndirectFitDataPresenter::updateDataInTable() { m_tablePresenter->updateTableFromModel(); }
-
-DataForParameterEstimationCollection
-IndirectFitDataPresenter::getDataForParameterEstimation(const EstimationDataSelector &selector) const {
-  return m_model->getDataForParameterEstimation(std::move(selector));
-}
 
 UserInputValidator &IndirectFitDataPresenter::validate(UserInputValidator &validator) {
   return m_view->validate(validator);
@@ -101,15 +99,78 @@ void IndirectFitDataPresenter::closeDialog() {
 void IndirectFitDataPresenter::addData(IAddWorkspaceDialog const *dialog) {
   try {
     emit dataAdded(dialog);
-    m_tablePresenter->updateTableFromModel();
+    updateTableFromModel();
     emit dataChanged();
   } catch (const std::runtime_error &ex) {
     displayWarning(ex.what());
   }
 }
 
+void IndirectFitDataPresenter::updateTableFromModel() {
+  ScopedFalse _signalBlock(m_emitCellChanged);
+  m_view->clearTable();
+  for (auto domainIndex = FitDomainIndex{0}; domainIndex < m_model->getNumberOfDomains(); domainIndex++) {
+    addTableEntry(domainIndex);
+  }
+}
 void IndirectFitDataPresenter::displayWarning(const std::string &warning) { m_view->displayWarning(warning); }
 
+void IndirectFitDataPresenter::addTableEntry(FitDomainIndex row) {
+  const auto &name = m_model->getWorkspace(row)->getName();
+  const auto workspaceIndex = m_model->getSpectrum(row);
+  const auto range = m_model->getFittingRange(row);
+  const auto exclude = m_model->getExcludeRegion(row);
+
+  FitDataRow newRow;
+  newRow.name = name;
+  newRow.workspaceIndex = workspaceIndex;
+  newRow.startX = range.first;
+  newRow.endX = range.second;
+  newRow.exclude = exclude;
+
+  m_view->addTableEntry(row.value, newRow);
+}
+
+void IndirectFitDataPresenter::handleCellChanged(int row, int column) {
+  if (!m_emitCellChanged) {
+    return;
+  }
+
+  if (m_view->startXColumn() == column) {
+    setModelStartXAndEmit(m_view->getText(row, column).toDouble(), row);
+  } else if (m_view->endXColumn() == column) {
+    setModelEndXAndEmit(m_view->getText(row, column).toDouble(), row);
+  } else if (m_view->excludeColumn() == column) {
+    setModelExcludeAndEmit(m_view->getText(row, column).toStdString(), row);
+  }
+}
+
+void IndirectFitDataPresenter::setModelStartXAndEmit(double startX, FitDomainIndex row) {
+  auto subIndices = m_model->getSubIndices(row);
+  m_model->setStartX(startX, subIndices.first, subIndices.second);
+  emit startXChanged(startX, subIndices.first, subIndices.second);
+}
+
+void IndirectFitDataPresenter::setModelEndXAndEmit(double endX, FitDomainIndex row) {
+  auto subIndices = m_model->getSubIndices(row);
+  m_model->setEndX(endX, subIndices.first, subIndices.second);
+  emit endXChanged(endX, subIndices.first, subIndices.second);
+}
+
+void IndirectFitDataPresenter::setModelExcludeAndEmit(const std::string &exclude, FitDomainIndex row) {
+  auto subIndices = m_model->getSubIndices(row);
+  m_model->setExcludeRegion(exclude, subIndices.first, subIndices.second);
+}
+
+void IndirectFitDataPresenter::removeSelectedData() {
+  auto selectedIndices = m_view->getSelectedIndexes();
+  std::sort(selectedIndices.begin(), selectedIndices.end());
+  for (auto item = selectedIndices.end(); item != selectedIndices.begin();) {
+    --item;
+    m_model->removeDataByIndex(FitDomainIndex(item->row()));
+  }
+  updateTableFromModel();
+}
 } // namespace IDA
 } // namespace CustomInterfaces
 } // namespace MantidQt
