+++ qt/scientific_interfaces/Indirect/IndirectFitDataView.h
@@ -11,6 +11,7 @@
 #include "IIndirectFitDataView.h"
 
 #include "DllConfig.h"
+#include "MantidQtWidgets/Common/IndexTypes.h"
 #include "MantidQtWidgets/Common/UserInputValidator.h"
 
 #include <QTabWidget>
@@ -26,14 +27,28 @@ public:
   ~IndirectFitDataView() override = default;
 
   QTableWidget *getDataTable() const override;
+  bool isTableEmpty() const;
 
   UserInputValidator &validate(UserInputValidator &validator) override;
+  virtual void addTableEntry(size_t row, FitDataRow newRow) override;
+  virtual int workspaceIndexColumn() const override;
+  virtual int startXColumn() const override;
+  virtual int endXColumn() const override;
+  virtual int excludeColumn() const override;
+  void clearTable() override;
+  QString getText(int row, int column) const override;
+  QModelIndexList getSelectedIndexes() const override;
 
 public slots:
   void displayWarning(const std::string &warning) override;
 
+protected:
+  IndirectFitDataView(const QStringList &headers, QWidget *parent = nullptr);
+  std::unique_ptr<Ui::IndirectFitDataView> m_uiForm;
+  void setCell(std::unique_ptr<QTableWidgetItem> cell, size_t row, size_t column);
+
 private:
-  std::unique_ptr<Ui::IndirectFitDataView> m_dataForm;
+  void setHorizontalHeaders(const QStringList &headers);
 };
 
 } // namespace IDA
