+++ qt/scientific_interfaces/Indirect/test/ConvFitDataPresenterTest.h
@@ -46,13 +46,64 @@ public:
   /// Public Methods
   MOCK_CONST_METHOD0(getDataTable, QTableWidget *());
   MOCK_METHOD1(validate, UserInputValidator &(UserInputValidator &validator));
+  MOCK_METHOD2(addTableEntry, void(size_t row, FitDataRow newRow));
+  MOCK_CONST_METHOD0(workspaceIndexColumn, int());
+  MOCK_CONST_METHOD0(startXColumn, int());
+  MOCK_CONST_METHOD0(endXColumn, int());
+  MOCK_CONST_METHOD0(excludeColumn, int());
+  MOCK_METHOD0(clearTable, void());
+  MOCK_CONST_METHOD2(getText, QString(int row, int column));
+  MOCK_CONST_METHOD0(getSelectedIndexes, QModelIndexList());
 
   /// Public slots
   MOCK_METHOD1(displayWarning, void(std::string const &warning));
 };
 
 /// Mock object to mock the model
-class MockConvFitModel : public ConvFitModel {};
+class MockIndirectConvFitDataModel : public IIndirectFitDataModel {
+public:
+  /// Public Methods
+  MOCK_CONST_METHOD1(hasWorkspace, bool(std::string const &workspaceName));
+  MOCK_CONST_METHOD1(getWorkspace, MatrixWorkspace_sptr(WorkspaceID workspaceID));
+  MOCK_CONST_METHOD1(getSpectra, FunctionModelSpectra(WorkspaceID workspaceID));
+  MOCK_CONST_METHOD0(isMultiFit, bool());
+  MOCK_CONST_METHOD0(getNumberOfWorkspaces, WorkspaceID());
+  MOCK_CONST_METHOD1(getNumberOfSpectra, size_t(WorkspaceID workspaceID));
+  MOCK_CONST_METHOD0(getNumberOfDomains, size_t());
+  MOCK_CONST_METHOD2(getDomainIndex, FitDomainIndex(WorkspaceID workspaceID, WorkspaceIndex spectrum));
+  MOCK_CONST_METHOD0(getQValuesForData, std::vector<double>());
+  MOCK_CONST_METHOD0(getResolutionsForFit, std::vector<std::pair<std::string, size_t>>());
+  MOCK_CONST_METHOD0(getWorkspaceNames, std::vector<std::string>());
+
+  MOCK_METHOD2(setSpectra, void(const std::string &spectra, WorkspaceID workspaceID));
+  MOCK_METHOD2(setSpectra, void(FunctionModelSpectra &&spectra, WorkspaceID workspaceID));
+  MOCK_METHOD2(setSpectra, void(const FunctionModelSpectra &spectra, WorkspaceID workspaceID));
+  MOCK_METHOD1(addWorkspace, void(const std::string &workspaceName));
+  MOCK_METHOD2(addWorkspace, void(const std::string &workspaceName, const std::string &spectra));
+  MOCK_METHOD2(addWorkspace, void(const std::string &workspaceName, const FunctionModelSpectra &spectra));
+  MOCK_METHOD2(addWorkspace, void(MatrixWorkspace_sptr workspace, const FunctionModelSpectra &spectra));
+  MOCK_METHOD1(removeWorkspace, void(WorkspaceID workspaceID));
+  MOCK_METHOD1(removeDataByIndex, void(FitDomainIndex fitDomainIndex));
+  MOCK_METHOD0(clear, void());
+
+  MOCK_CONST_METHOD2(getFittingRange, std::pair<double, double>(WorkspaceID workspaceID, WorkspaceIndex spectrum));
+  MOCK_CONST_METHOD2(getExcludeRegion, std::string(WorkspaceID workspaceID, WorkspaceIndex spectrum));
+  MOCK_CONST_METHOD2(getExcludeRegionVector, std::vector<double>(WorkspaceID workspaceID, WorkspaceIndex spectrum));
+  MOCK_METHOD3(setStartX, void(double startX, WorkspaceID workspaceID, WorkspaceIndex spectrum));
+  MOCK_METHOD2(setStartX, void(double startX, WorkspaceID workspaceID));
+  MOCK_METHOD3(setEndX, void(double endX, WorkspaceID workspaceID, WorkspaceIndex spectrum));
+  MOCK_METHOD2(setEndX, void(double endX, WorkspaceID workspaceID));
+  MOCK_METHOD3(setExcludeRegion, void(const std::string &exclude, WorkspaceID workspaceID, WorkspaceIndex spectrum));
+  MOCK_METHOD2(setResolution, void(const std::string &name, WorkspaceID workspaceID));
+
+  MOCK_CONST_METHOD1(getWorkspace, MatrixWorkspace_sptr(FitDomainIndex index));
+  MOCK_CONST_METHOD1(getFittingRange, std::pair<double, double>(FitDomainIndex index));
+  MOCK_CONST_METHOD1(getSpectrum, size_t(FitDomainIndex index));
+  MOCK_CONST_METHOD1(getExcludeRegionVector, std::vector<double>(FitDomainIndex index));
+  MOCK_CONST_METHOD1(getExcludeRegion, std::string(FitDomainIndex index));
+  MOCK_METHOD2(setExcludeRegion, void(const std::string &exclude, FitDomainIndex index));
+  MOCK_CONST_METHOD1(getSubIndices, std::pair<WorkspaceID, WorkspaceIndex>(FitDomainIndex));
+};
 
 GNU_DIAG_ON_SUGGEST_OVERRIDE
 
@@ -67,7 +118,7 @@ public:
 
   void setUp() override {
     m_view = std::make_unique<NiceMock<MockConvFitDataView>>();
-    m_model = std::make_unique<NiceMock<MockConvFitModel>>();
+    m_model = std::make_unique<NiceMock<MockIndirectConvFitDataModel>>();
 
     m_dataTable = createEmptyTableWidget(6, 5);
     ON_CALL(*m_view, getDataTable()).WillByDefault(Return(m_dataTable.get()));
@@ -117,6 +168,6 @@ private:
   std::unique_ptr<QTableWidget> m_dataTable;
 
   std::unique_ptr<MockConvFitDataView> m_view;
-  std::unique_ptr<MockConvFitModel> m_model;
+  std::unique_ptr<MockIndirectConvFitDataModel> m_model;
   std::unique_ptr<ConvFitDataPresenter> m_presenter;
 };
