+++ qt/scientific_interfaces/Indirect/test/IndirectFitDataPresenterTest.h
@@ -10,9 +10,8 @@
 #include <gmock/gmock.h>
 
 #include "IIndirectFitDataView.h"
+#include "IndirectFitDataModel.h"
 #include "IndirectFitDataPresenter.h"
-#include "IndirectFitDataTableModel.h"
-#include "IndirectFitDataTablePresenterTest.h"
 #include "IndirectFitDataView.h"
 #include "IndirectFittingModel.h"
 #include "ParameterEstimation.h"
@@ -27,6 +26,32 @@ using namespace MantidQt::CustomInterfaces;
 using namespace MantidQt::CustomInterfaces::IDA;
 using namespace testing;
 
+namespace {
+
+std::unique_ptr<QTableWidget> createEmptyTableWidget(int columns, int rows) {
+  auto table = std::make_unique<QTableWidget>(columns, rows);
+  for (auto column = 0; column < columns; ++column)
+    for (auto row = 0; row < rows; ++row)
+      table->setItem(row, column, new QTableWidgetItem("item"));
+  return table;
+}
+
+struct TableItem {
+  TableItem(std::string const &value) : m_str(value), m_dbl(0.0) {}
+  TableItem(double const &value) : m_str(QString::number(value, 'g', 16).toStdString()), m_dbl(value) {}
+
+  std::string const &asString() const { return m_str; }
+  QString asQString() const { return QString::fromStdString(m_str); }
+  double const &asDouble() const { return m_dbl; }
+
+  bool operator==(std::string const &value) const { return this->asString() == value; }
+
+private:
+  std::string m_str;
+  double m_dbl;
+};
+} // namespace
+
 GNU_DIAG_OFF_SUGGEST_OVERRIDE
 
 /// Mock object to mock the view
@@ -35,30 +60,23 @@ public:
   /// Public Methods
   MOCK_CONST_METHOD0(getDataTable, QTableWidget *());
   MOCK_METHOD1(validate, UserInputValidator &(UserInputValidator &validator));
+  MOCK_METHOD2(addTableEntry, void(size_t row, FitDataRow newRow));
+  MOCK_CONST_METHOD0(workspaceIndexColumn, int());
+  MOCK_CONST_METHOD0(startXColumn, int());
+  MOCK_CONST_METHOD0(endXColumn, int());
+  MOCK_CONST_METHOD0(excludeColumn, int());
+  MOCK_METHOD0(clearTable, void());
+  MOCK_CONST_METHOD2(getText, QString(int row, int column));
+  MOCK_CONST_METHOD0(getSelectedIndexes, QModelIndexList());
 
   /// Public slots
   MOCK_METHOD1(displayWarning, void(std::string const &warning));
 };
 
 /// Mock object to mock the model
-class MockIndirectFitDataTableModel : public IIndirectFittingModel {
+class MockIndirectFitDataModel : public IIndirectFitDataModel {
 public:
-  MockIndirectFitDataTableModel() : m_fitDataModel(std::make_unique<MockIndirectDataTableModel>()) {}
-
   /// Public Methods
-  MOCK_CONST_METHOD2(isPreviouslyFit, bool(WorkspaceID workspaceID, WorkspaceIndex spectrum));
-  MOCK_CONST_METHOD0(isInvalidFunction, boost::optional<std::string>());
-  MOCK_CONST_METHOD0(getFitParameterNames, std::vector<std::string>());
-  MOCK_CONST_METHOD0(getFitFunction, Mantid::API::MultiDomainFunction_sptr());
-  MOCK_CONST_METHOD2(getParameterValues, std::unordered_map<std::string, IDA::ParameterValue>(WorkspaceID workspaceID,
-                                                                                              WorkspaceIndex spectrum));
-  MOCK_METHOD1(setFitFunction, void(Mantid::API::MultiDomainFunction_sptr function));
-  MOCK_METHOD3(setDefaultParameterValue, void(const std::string &name, double value, IDA::WorkspaceID workspaceID));
-
-  MOCK_CONST_METHOD2(getFitParameters, std::unordered_map<std::string, IDA::ParameterValue>(WorkspaceID workspaceID,
-                                                                                            WorkspaceIndex spectrum));
-  MOCK_CONST_METHOD1(getDefaultParameters,
-                     std::unordered_map<std::string, IDA::ParameterValue>(WorkspaceID workspaceID));
   MOCK_CONST_METHOD1(hasWorkspace, bool(std::string const &workspaceName));
   MOCK_CONST_METHOD1(getWorkspace, MatrixWorkspace_sptr(WorkspaceID workspaceID));
   MOCK_CONST_METHOD1(getSpectra, FunctionModelSpectra(WorkspaceID workspaceID));
@@ -69,67 +87,36 @@ public:
   MOCK_CONST_METHOD2(getDomainIndex, FitDomainIndex(WorkspaceID workspaceID, WorkspaceIndex spectrum));
   MOCK_CONST_METHOD0(getQValuesForData, std::vector<double>());
   MOCK_CONST_METHOD0(getResolutionsForFit, std::vector<std::pair<std::string, size_t>>());
-  MOCK_METHOD0(clearWorkspaces, void());
-  MOCK_METHOD0(clear, void());
+  MOCK_CONST_METHOD0(getWorkspaceNames, std::vector<std::string>());
+
   MOCK_METHOD2(setSpectra, void(const std::string &spectra, WorkspaceID workspaceID));
   MOCK_METHOD2(setSpectra, void(FunctionModelSpectra &&spectra, WorkspaceID workspaceID));
   MOCK_METHOD2(setSpectra, void(const FunctionModelSpectra &spectra, WorkspaceID workspaceID));
-
-  MOCK_METHOD1(addWorkspace, void(std::string const &workspaceName));
-  MOCK_METHOD2(addWorkspace, void(std::string const &workspaceName, std::string const &spectra));
-  MOCK_METHOD2(addWorkspace, void(std::string const &workspaceName, FunctionModelSpectra const &spectra));
-  MOCK_METHOD2(addWorkspace, void(Mantid::API::MatrixWorkspace_sptr workspace, FunctionModelSpectra const &spectra));
+  MOCK_METHOD1(addWorkspace, void(const std::string &workspaceName));
+  MOCK_METHOD2(addWorkspace, void(const std::string &workspaceName, const std::string &spectra));
+  MOCK_METHOD2(addWorkspace, void(const std::string &workspaceName, const FunctionModelSpectra &spectra));
+  MOCK_METHOD2(addWorkspace, void(MatrixWorkspace_sptr workspace, const FunctionModelSpectra &spectra));
   MOCK_METHOD1(removeWorkspace, void(WorkspaceID workspaceID));
+  MOCK_METHOD1(removeDataByIndex, void(FitDomainIndex fitDomainIndex));
+  MOCK_METHOD0(clear, void());
 
   MOCK_CONST_METHOD2(getFittingRange, std::pair<double, double>(WorkspaceID workspaceID, WorkspaceIndex spectrum));
   MOCK_CONST_METHOD2(getExcludeRegion, std::string(WorkspaceID workspaceID, WorkspaceIndex spectrum));
+  MOCK_CONST_METHOD2(getExcludeRegionVector, std::vector<double>(WorkspaceID workspaceID, WorkspaceIndex spectrum));
   MOCK_METHOD3(setStartX, void(double startX, WorkspaceID workspaceID, WorkspaceIndex spectrum));
   MOCK_METHOD2(setStartX, void(double startX, WorkspaceID workspaceID));
   MOCK_METHOD3(setEndX, void(double endX, WorkspaceID workspaceID, WorkspaceIndex spectrum));
   MOCK_METHOD2(setEndX, void(double endX, WorkspaceID workspaceID));
   MOCK_METHOD3(setExcludeRegion, void(const std::string &exclude, WorkspaceID workspaceID, WorkspaceIndex spectrum));
+  MOCK_METHOD2(setResolution, void(const std::string &name, WorkspaceID workspaceID));
 
-  MOCK_METHOD3(addSingleFitOutput, void(const Mantid::API::IAlgorithm_sptr &fitAlgorithm, WorkspaceID workspaceID,
-                                        WorkspaceIndex spectrum));
-  MOCK_METHOD1(addOutput, void(Mantid::API::IAlgorithm_sptr fitAlgorithm));
-
-  MOCK_METHOD0(switchToSingleInputMode, void());
-  MOCK_METHOD0(switchToMultipleInputMode, void());
-  MOCK_METHOD1(setFittingMode, void(FittingMode mode));
-  MOCK_CONST_METHOD0(getFittingMode, FittingMode());
-  MOCK_METHOD1(setFitTypeString, void(const std::string &fitType));
-  MOCK_CONST_METHOD2(getResultLocation,
-                     boost::optional<ResultLocationNew>(WorkspaceID workspaceID, WorkspaceIndex spectrum));
-
-  MOCK_CONST_METHOD0(getResultWorkspace, Mantid::API::WorkspaceGroup_sptr());
-  MOCK_CONST_METHOD0(getResultGroup, Mantid::API::WorkspaceGroup_sptr());
-  MOCK_CONST_METHOD0(getFittingAlgorithm, Mantid::API::IAlgorithm_sptr());
-  MOCK_CONST_METHOD2(getSingleFit, Mantid::API::IAlgorithm_sptr(WorkspaceID workspaceID, WorkspaceIndex spectrum));
-  MOCK_CONST_METHOD2(getSingleFunction, Mantid::API::IFunction_sptr(WorkspaceID workspaceID, WorkspaceIndex spectrum));
-
-  MOCK_CONST_METHOD0(getOutputBasename, std::string());
-  MOCK_CONST_METHOD1(createDisplayName, std::string(WorkspaceID workspaceID));
-  MOCK_METHOD1(cleanFailedRun, void(const Mantid::API::IAlgorithm_sptr &fittingAlgorithm));
-  MOCK_METHOD2(cleanFailedSingleRun,
-               void(const Mantid::API::IAlgorithm_sptr &fittingAlgorithm, WorkspaceID workspaceID));
-  MOCK_CONST_METHOD1(getDataForParameterEstimation,
-                     DataForParameterEstimationCollection(const EstimationDataSelector &selector));
-
-  MOCK_METHOD0(removeFittingData, void());
-
-  IIndirectFitDataTableModel *getFitDataModel() override { return m_fitDataModel.get(); }
-
-protected:
-  std::unique_ptr<IIndirectFitDataTableModel> m_fitDataModel;
-
-private:
-  std::string sequentialFitOutputName() const { return ""; };
-  std::string simultaneousFitOutputName() const { return ""; };
-  std::string singleFitOutputName(WorkspaceID workspaceID, WorkspaceIndex spectrum) const {
-    UNUSED_ARG(workspaceID);
-    UNUSED_ARG(spectrum);
-    return "";
-  };
+  MOCK_CONST_METHOD1(getWorkspace, MatrixWorkspace_sptr(FitDomainIndex index));
+  MOCK_CONST_METHOD1(getFittingRange, std::pair<double, double>(FitDomainIndex index));
+  MOCK_CONST_METHOD1(getSpectrum, size_t(FitDomainIndex index));
+  MOCK_CONST_METHOD1(getExcludeRegionVector, std::vector<double>(FitDomainIndex index));
+  MOCK_CONST_METHOD1(getExcludeRegion, std::string(FitDomainIndex index));
+  MOCK_METHOD2(setExcludeRegion, void(const std::string &exclude, FitDomainIndex index));
+  MOCK_CONST_METHOD1(getSubIndices, std::pair<WorkspaceID, WorkspaceIndex>(FitDomainIndex));
 };
 MATCHER_P(NoCheck, selector, "") { return arg != selector; }
 
@@ -173,12 +160,11 @@ public:
 
   void setUp() override {
     m_view = std::make_unique<NiceMock<MockIIndirectFitDataView>>();
-    m_model = std::make_unique<NiceMock<MockIndirectFitDataTableModel>>();
+    m_model = std::make_unique<NiceMock<MockIndirectFitDataModel>>();
     m_table = createEmptyTableWidget(5, 5);
     ON_CALL(*m_view, getDataTable()).WillByDefault(Return(m_table.get()));
     m_presenter = std::make_unique<IndirectFitDataPresenter>(std::move(m_model.get()), std::move(m_view.get()));
 
-    TS_ASSERT(m_model->getFitDataModel());
     m_ads = std::make_unique<SetUpADSWithWorkspace>("WorkspaceName", createWorkspace(5));
     m_model->addWorkspace("WorkspaceName");
   }
@@ -214,32 +200,6 @@ public:
     TS_ASSERT_EQUALS(m_presenter->getSampleWSSuffices(), suffices);
   }
 
-  void test_that_setSampleFBSuffices_will_set_the_sample_file_browser_suffices_in_the_view() {
-    QStringList const suffices{"suffix1", "suffix2"};
-    m_presenter->setSampleFBSuffices(suffices);
-    TS_ASSERT_EQUALS(m_presenter->getSampleFBSuffices(), suffices);
-  }
-
-  void test_that_setResolutionWSSuffices_will_set_the_resolution_workspace_suffices_in_the_view() {
-    QStringList const suffices{"suffix1", "suffix2"};
-    m_presenter->setResolutionWSSuffices(suffices);
-    TS_ASSERT_EQUALS(m_presenter->getResolutionWSSuffices(), suffices);
-  }
-
-  void test_that_setResolutionFBSuffices_will_set_the_resolution_file_browser_suffices_in_the_view() {
-    QStringList const suffices{"suffix1", "suffix2"};
-    m_presenter->setResolutionFBSuffices(suffices);
-    TS_ASSERT_EQUALS(m_presenter->getResolutionFBSuffices(), suffices);
-  }
-
-  void test_getDataForParameterEstimation_uses_selector_to_get_from_model() {
-    EstimationDataSelector selector = getEstimationDataSelector();
-
-    EXPECT_CALL(*m_model, getDataForParameterEstimation(NoCheck(nullptr))).Times(Exactly(1));
-
-    m_presenter->getDataForParameterEstimation(selector);
-  }
-
 private:
   void deleteSetup() {
     m_presenter.reset();
@@ -259,7 +219,7 @@ private:
   std::unique_ptr<QTableWidget> m_table;
 
   std::unique_ptr<MockIIndirectFitDataView> m_view;
-  std::unique_ptr<MockIndirectFitDataTableModel> m_model;
+  std::unique_ptr<MockIndirectFitDataModel> m_model;
   std::unique_ptr<IndirectFitDataPresenter> m_presenter;
 
   std::unique_ptr<SetUpADSWithWorkspace> m_ads;
